增加历史纪录  和 退出删除的接口

+++ FactoryBuildingOnline/Main(主要)/Me(我的)/Controller/UserInfomationViewController.m
@@ -196,8 +196,20 @@
             
         default:    // 退出登录
         {
-            [FOLUserInforModel deleteAll];
-            [self.navigationController popViewControllerAnimated:YES];
+            [HTTPREQUEST_SINGLE delectWithQuitLogin:URL_DELECT_QUIT andParameters:nil isShowActivity:YES success:^(RequestManager *manager, NSDictionary *response) {
+                
+                NSLog(@"用户退出：%@",response);
+                
+                if ([response[@"erro_code"] intValue] != 200) {
+                    return ;
+                }
+                
+                [FOLUserInforModel deleteAll];
+                [self.navigationController popViewControllerAnimated:YES];
+            } failure:^(RequestManager *manager, NSError *error) {
+                NSLog(@"用户退出：%@",error);
+            }];
+            
         }
             break;
     }
